Tempus-1073 Saving files to S3 with white spaces.

+++ application/src/main/java/com/hashmapinc/server/service/entityfile/EntityFileService.java
@@ -83,9 +83,8 @@ public class EntityFileService {
         String bucketName = CloudStorageServiceUtils.createBucketName(tenantService.findTenantById(tenantId));
         String oldObjectUrl = CloudStorageServiceUtils.createObjectName(oldFileName, entityId, StorageTypes.FILES);
         String newObjectUrl = CloudStorageServiceUtils.createObjectName(newFileName, entityId, StorageTypes.FILES);
-        if (cloudStorageService.copyFile(bucketName, oldObjectUrl, newObjectUrl)) {
-            if(!oldObjectUrl.contentEquals(newObjectUrl))
-                cloudStorageService.delete(bucketName, oldObjectUrl);
+        if (!oldObjectUrl.contentEquals(newObjectUrl) && cloudStorageService.copyFile(bucketName, oldObjectUrl, newObjectUrl)) {
+            cloudStorageService.delete(bucketName, oldObjectUrl);
             renameFileInDb(oldFileName, newFileName, tenantId, entityId);
         }
     }
@@ -117,9 +116,9 @@ public class EntityFileService {
     private String getFileNameWithOrWithoutExt(String fileName) {
         String[] arrList = fileName.split("\\.");
         if (arrList.length >= 2) {
-            return fileName.substring(0, fileName.lastIndexOf(".")).replace(" ", "-");
+            return fileName.substring(0, fileName.lastIndexOf("."));
         }
-        return fileName.replace(" ", "-");
+        return fileName;
     }
 
     private String getFileExtension(String fileName) {
